rm msvc warning

+++ src/app/midiProvider.h
@@ -24,7 +24,7 @@ namespace app
 			NO_TICK = INT_MAX,
 			INVALID_TRACKID = 0
 		};
-		typedef long TrackId;
+		typedef intptr_t TrackId;
 		typedef long double Millis;
 		struct Event
 		{
